2.5.2.2.0006: Fixed Issue #2 [NumberFormatException for precipitation field.]

+++ src/main/kotlin/net/aksingh/owmjapis/model/param/Cloud.kt
@@ -24,7 +24,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Cloud(
   @field:SerializedName("all")
-  val cloud: Int? = null
+  val cloud: Double? = null
 ) {
 
   fun hasCloud(): Boolean = cloud != null
